refactor: separate the modal content in a separate component
