+++ src/components/TranslationDialog/TranslationModal/TranslationModal.js
@@ -5,22 +5,14 @@ import {
     ButtonStrip,
     CenteredContent,
     CircularLoader,
-    DataTable,
-    DataTableBody,
-    DataTableCell,
-    DataTableColumnHeader,
-    DataTableHead,
-    DataTableRow,
-    Input,
     Modal,
     ModalActions,
     ModalContent,
     ModalTitle,
 } from '@dhis2/ui'
-import cloneDeep from 'lodash/cloneDeep'
 import PropTypes from 'prop-types'
 import React, { useEffect, useMemo, useState } from 'react'
-import { LocalesSelect } from './LocalesSelect.js'
+import { TranslationForm } from './TranslationForm.js'
 import { useTranslationsResults } from './useTranslationsResults.js'
 
 const getTranslationsMutation = resource => ({
@@ -38,7 +30,6 @@ export const TranslationModal = ({
     const { show: showError } = useAlert(error => error, { critical: true })
 
     const [translations, setTranslations] = useState([])
-    const [translationLocale, setTranslationLocale] = useState()
 
     const endpointPath = new URL(objectToTranslate.href).pathname
     const endpointPathMatch = endpointPath.match(/api\/\d+\/(?<resource>.+)/)
@@ -69,51 +60,6 @@ export const TranslationModal = ({
 
     const save = () => updateTranslations({ translations })
 
-    const camelCaseToUnderscores = field =>
-        field
-            .replace(/[a-z][A-Z]/g, match =>
-                [match.chartAt(0), match.chartAt(1)].join('_')
-            )
-            .toLowerCase()
-
-    const getTranslationIndexForField = field => {
-        console.log('translations', translations, camelCaseToUnderscores(field))
-        return translations.findIndex(
-            element =>
-                element.locale === translationLocale &&
-                element.property.toLowerCase() === camelCaseToUnderscores(field)
-        )
-    }
-
-    const getTranslationForField = field => {
-        const translationIndex = getTranslationIndexForField(field)
-        console.log(field, translationIndex)
-        return translationIndex !== -1
-            ? translations[translationIndex]?.value || ''
-            : ''
-    }
-
-    const setTranslationForField = (field, translation) => {
-        console.log('setTranslationField called', field, translation)
-        const newTranslations = cloneDeep(translations)
-
-        const newTranslation = {
-            locale: translationLocale,
-            property: camelCaseToUnderscores(field).toUpperCase(),
-            value: translation,
-        }
-
-        const translationIndex = getTranslationIndexForField(field)
-
-        if (translationIndex !== -1) {
-            newTranslations.splice(translationIndex, 1, newTranslation)
-        } else {
-            newTranslations.push(newTranslation)
-        }
-
-        setTranslations(newTranslations)
-    }
-
     useEffect(() => {
         if (translationsData) {
             setTranslations(translationsData)
@@ -129,71 +75,19 @@ export const TranslationModal = ({
                 })}
             </ModalTitle>
             <ModalContent>
-                <DataTable layout="fixed">
-                    <DataTableHead>
-                        <DataTableRow>
-                            <DataTableColumnHeader fixed top="0">
-                                {i18n.t('Base locale reference')}
-                            </DataTableColumnHeader>
-                            <DataTableColumnHeader fixed top="0">
-                                <LocalesSelect
-                                    selected={translationLocale}
-                                    onChange={setTranslationLocale}
-                                />
-                            </DataTableColumnHeader>
-                        </DataTableRow>
-                    </DataTableHead>
-                    <DataTableBody>
-                        {fieldsToTranslate.map((field, index) => (
-                            <DataTableRow key={field}>
-                                <DataTableCell>
-                                    <div className="">
-                                        {field}
-                                        <Input
-                                            value={objectToTranslate[field]}
-                                            readOnly
-                                        />
-                                    </div>
-                                </DataTableCell>
-                                {translationLocale && (
-                                    <DataTableCell>
-                                        <div className="">
-                                            {field}
-                                            <Input
-                                                value={getTranslationForField(
-                                                    field
-                                                )}
-                                                onChange={({ value }) =>
-                                                    setTranslationForField(
-                                                        field,
-                                                        value
-                                                    )
-                                                }
-                                            />
-                                        </div>
-                                    </DataTableCell>
-                                )}
-                                {!translationLocale && index === 0 && (
-                                    <DataTableCell
-                                        rowSpan={String(
-                                            fieldsToTranslate.length
-                                        )}
-                                    >
-                                        <CenteredContent>
-                                            {fetching ? (
-                                                <CircularLoader />
-                                            ) : (
-                                                i18n.t(
-                                                    'Choose a locale to translate from the menu above'
-                                                )
-                                            )}
-                                        </CenteredContent>
-                                    </DataTableCell>
-                                )}
-                            </DataTableRow>
-                        ))}
-                    </DataTableBody>
-                </DataTable>
+                {fetching && (
+                    <CenteredContent>
+                        <CircularLoader />
+                    </CenteredContent>
+                )}
+                {translations.length && (
+                    <TranslationForm
+                        fieldsToTranslate={fieldsToTranslate}
+                        objectToTranslate={objectToTranslate}
+                        translations={translations}
+                        onUpdateTranslations={setTranslations}
+                    />
+                )}
             </ModalContent>
             <ModalActions>
                 <ButtonStrip>
